Agents

Agents

RLCore.get_action — Method.

get_action(agent::AbstractAgent, state, rng::AbstractRNG; kwargs...)

source

RLCore.get_action — Method.

get_action(agent::AbstractAgent, state; kwargs...)

Returns an action, deterministacally.

source

RLCore.start! — Method.

start!(agent::AbstractAgent, env_s_tp1, rng=Random.GLOBAL_RNG, kwargs...)

source

RLCore.start! — Method.

start!(agent::AbstractAgent, env_s_tp1, rng=Random.GLOBAL_RNG, kwargs...)

Function for starting the agent for a new episode.

returns an action to get passed to the environment

source

RLCore.step! — Method.

step!(agent::AbstractAgent, env_s_tp1, r, terminal, rng::AbstractRNG; kwargs...)

source

RLCore.step! — Method.

step!(agent::AbstractAgent, env_s_tp1, r, terminal; kwargs...)

Function to take a step with an agent.

Returns an action to get passed to the environment.

source