Feature Constructor

Feature Constructors

RLCore.AbstractFeatureConstructor — Type.

AbstractFeatureCreator

An abstract feature creator, for feature transformation from the states.

source

RLCore.create_features — Method.

create_features(fc::AbstractFeatureCreator, s, a)

Actually create the features using actions and states

source

RLCore.create_features — Method.

create_features(fc::AbstractFeatureCreator, s)

Actually create the features using only states

source

RLCore.feature_size — Method.

feature_size(fc::AbstractFeatureCreator)

Get size of feature vector the features assume exists.

source