GVF

GVFs

RLCore.GVF — Type.

GVF{C<:AbstractCumulant, D<:AbstractDiscount, P<:AbstractPolicy} <: AbstractGVF

A realized version of a GVF where the cumulant, discount, and policies can be any structure following the AbstractCumulant, AbstractDiscount, or AbstractPolicy api respectively.

source

RLCore.Horde — Type.

Horde{T<:AbstractGVF} <: AbstractHorde

The simplest implementation of a horde as a collection of AbstractGVFs.

source

RLCore.AbstractGVF — Type.

AbstractGVF

This is the base type for a General Value Function. See the "Horde: A scalable real-time architecture for learning knowledge from unsupervised sensorimotor interaction" paper for more details.

source

RLCore.AbstractHorde — Type.

AbstractHorde

An abstract collection of GVFs.

source

Base.get — Method.

get(gvf::AbstractGVF, state_t, action_t, state_tp1, action_tp1, preds_tp1)

Get the parameters for the cumulant, discount, and probability of taking an action given the parameters.

source

Base.get — Method.

get(gvf::AbstractGVF, state_t, action_t, state_tp1, preds_tp1)

Convenience function: get(gvf, state_t, action_t, state_tp1, nothing, preds_tp1)

source

Base.get — Method.

get(gvf::AbstractGVF, state_t, action_t, state_tp1)

Convenience function: get(gvf, state_t, action_t, state_tp1, nothing, nothing)

source